basic game playing structures added
without interfaces ruby makes this interesting, but have taken the view that a Neural net shouldn't be a player, instead Deep Beige currently chooses to create neural nets.  It may choose otherwise in the future.  We need to put error checking in of course to check that the games we add and player classes have the methods (else table will fail) but otherwise I quite like this approach.