changed sigmoid function to tanh to generate values between -1 and 1
this had the interesting side effect that since tanh(0) == 0 we needed to have the base 'best score' set to below -1.  It also bears thinking about whether, as we describe an empty square as having input value 0 this is a bit artificial.  It will instantly mean the neural net only recognises patterns of pieces and ignores patterns in white space.  This is probably good and eliminates much wasted effort in analysing empty space - but worth experimenting with - especially as patterns of white space are important in go